travis.yml

+++ src/test/scala/org/sparkRedshift/tutorial/test/AcceptanceTestCommonsTriggers.scala
@@ -3,6 +3,7 @@ package org.sparkRedshift.tutorial.test
 
 import cucumber.api.Scenario
 import cucumber.api.scala.{EN, ScalaDsl}
+import org.junit.Assert._
 import org.scalatest.Matchers
 import org.slf4j.{Logger, LoggerFactory}
 
@@ -17,8 +18,12 @@ class AcceptanceTestCommonsTriggers extends Matchers with ScalaDsl with EN with 
   }
 
   After("@cleanRedshiftRecords") { scenario: Scenario =>
+    if(clusterExist())
     deleteTable("category")
   }
 
-  def uuid() = java.util.UUID.randomUUID.toString
+  Before("@checkClusterItsUp") { scenario: Scenario =>
+    assertTrue(s"clusterDown!", clusterExist())
+  }
+
 }
